watchlist_app/api/v1/custom_throttle.py
from rest_framework.throttling import UserRateThrottle, AnonRateThrottle


class ReviewThrottle(UserRateThrottle):
    scope = 'review'

class AnonymousReviewThrottle(AnonRateThrottle):
    scope = 'review'